Move logger and types to core folder
