+++ webapp/src/core/logger.ts
@@ -1,6 +1,6 @@
 
 
-class Logger {
+export class Logger {
     log(message: unknown, ...args: unknown[]) {
         if (import.meta.env.VITE_DEV_LOGGER) {
             console.log(message, ...args);
