+++ webapp/src/core/types.ts
